backend/pyproject.toml
[project]
name = "rbac-mlflow"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.135.1",
    "uvicorn[standard]==0.42.0",
    "python-jose[cryptography]==3.5.0",
    "httpx==0.28.1",
    "pydantic-settings==2.9.1",
    "sqlalchemy[asyncio]==2.0.48",
    "asyncpg==0.31.0",
    "alembic==1.18.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rbac_mlflow"]

[tool.uv]
dev-dependencies = [
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "httpx==0.28.1",
    "ruff==0.15.6",
    "ty==0.0.23",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
